Move SGX specific header to appropriate subfolder
Move internal/ecall_context.h to internal/sgx/ecall_context.h
