+++ enclave/core/sgx/hostcalls.c
@@ -3,7 +3,7 @@
 
 #include <openenclave/edger8r/enclave.h>
 #include <openenclave/enclave.h>
-#include <openenclave/internal/ecall_context.h>
+#include <openenclave/internal/sgx/ecall_context.h>
 #include "td.h"
 
 /**
