+++ include/openenclave/internal/sgx/ecall_context.h
